Change Log 1.1.0
- Server-Server com now has a password that must match for servers to communicate
- You now need server password to use auth id instead of access token
- Routes now have icons corresponding to what they are
- loading screen is now dark and has a spinning wheel as well as large centered text
- users can now create their own categories
- some types are now moved to dedicated file
- some variables now have own file instead of constantly being redefined
- userData is renamed to userInfo
- new table for storing user settings(currently only categories)
- Stats screen global stats have been changed to Total tracked manga, Total tracked Chapters, and 2 TBD
- new page for viewing all user tracked manga and manage them
- new settings screen (currently only used for categories)
- verifying users has been moved to dedicated function that handles getting user ID or returning error message
- resolves changeMangaCat not always changing the category
- commented out code in service worker as it caused issue deploying and wasnt used.
- puppeteer server has been changed to use typescript
- resolved issue where when manga failed to pull during update none of the data was saved
- added option for verbose logging in puppeteer server
- updated asura pull utils for the new webpage
- created new util for converting old asura links to new ones
- utils now has config file instead of hard coded values

+++ mangaActivity/packages/server/src/dataUtils/removeUtils.ts
@@ -1,9 +1,12 @@
 import {Env, userDataRow, mangaDataRowReturn, user, mangaReturn} from '../types'
-import {getUserID} from '../utils'
+import {verifyUserAuth} from '../utils'
 
 export async function forgetUser(access_token:string|null=null, authId:string, env:Env) {
     try {
-        if (access_token) authId = await getUserID(access_token)
+        const validationRes = await verifyUserAuth(access_token, authId, env)
+        
+        if (validationRes instanceof Response) return validationRes
+        authId = validationRes
 
         await env.DB.prepare('DELETE FROM userData WHERE userID = ?')
                 .bind(authId)
@@ -11,13 +14,16 @@ export async function forgetUser(access_token:string|null=null, authId:string, e
         return new Response(JSON.stringify({message: "Success"}), {status:200})
     } catch (err) {
         console.error("Error:", err);
-        return new Response(JSON.stringify({message: 'an unknown error occured'}), {status:500});
+        return new Response(JSON.stringify({message: 'an unknown error occurred'}), {status:500});
     }
 }
 
 export async function deleteUserManga(access_token:string|null=null, authId:string, mangaId:string, env:Env) {
     try {
-        if (access_token) authId = await getUserID(access_token)
+        const validationRes = await verifyUserAuth(access_token, authId, env)
+        
+        if (validationRes instanceof Response) return validationRes
+        authId = validationRes
 
         await env.DB.prepare('DELETE FROM userData WHERE userID = ? AND mangaId = ?')
                 .bind(authId, mangaId)
@@ -25,6 +31,6 @@ export async function deleteUserManga(access_token:string|null=null, authId:stri
         return new Response(JSON.stringify({message: "Success", data: authId+mangaId}), {status:200})
     } catch (err) {
         console.error("Error:", err);
-        return new Response(JSON.stringify({message: 'an unknown error occured'}), {status:500});
+        return new Response(JSON.stringify({message: 'an unknown error occurred'}), {status:500});
     }
 }
